+++ NEWS.md
@@ -1,3 +1,11 @@
+# SplicingFactory 0.99.2
+
+* SummarizedExperiment input type instead of ExpressionSet.
+
+# SplicingFactory 0.99.1
+
+* Correction: unnecessary file removed.
+
 # SplicingFactory 0.99.0 (dev)
 
 * Submitted to Bioconductor